requirements.txt
google-generativeai
moviepy

